feat: implemented review aspects for copy commit feature
